# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transporte API",
        "description": "Documentación generada con Swagger para la API de transporte",
        "contact": {
            "email": "soporte@transporte.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        },
        {
            "url": "https://tu-dominio.com",
            "description": "Servidor Producción"
        }
    ],
    "paths": {
...
        "/api/tipos-mercancia": {
            "get": {
                "tags": [
                    "Tipos de Mercancía"
                ],
                "summary": "Listar tipos de mercancía",
                "description": "Obtiene un listado paginado de tipos de mercancía.",
                "operationId": "42a8525408d6911cbf3c75cbf20b3485",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de mercancía",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TipoMercancia"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tipos de Mercancía"
                ],
                "summary": "Crear un nuevo tipo de mercancía",
                "operationId": "c208cb7d3ffee38a0aff9478074b9265",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Electrónica"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de mercancía creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoMercancia"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tipos-mercancia/{id}": {
            "get": {
                "tags": [
                    "Tipos de Mercancía"
                ],
                "summary": "Mostrar un tipo de mercancía",
                "description": "Devuelve la información de un tipo de mercancía específico.",
                "operationId": "74712f5f9663b327e091b33122c6d210",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de mercancía",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de mercancía encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoMercancia"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de mercancía no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de mercancía no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tipos de Mercancía"
                ],
                "summary": "Actualizar un tipo de mercancía",
                "operationId": "8802bc6e2b662ea784868581f9bcde2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de mercancía",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Alimentos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de mercancía actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoMercancia"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tipos de Mercancía"
                ],
                "summary": "Eliminar un tipo de mercancía",
                "operationId": "94e54a6a0ec3c4da9b3a9e736b31a7c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo de mercancía",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de mercancía eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "TipoMercancia": {
                "title": "TipoMercancia",
                "description": "Catálogo de tipos de mercancías disponibles para los paquetes",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "tipo": {
                        "description": "Nombre o categoría del tipo de mercancía",
                        "type": "string",
                        "example": "Electrodomésticos"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}